+++ src/main/java/org/bytesync/hotelmanagement/dto/room/RoomDto.java
@@ -7,8 +7,8 @@ import org.bytesync.hotelmanagement.model.enums.RoomStatus;
 @Builder
 public record RoomDto(
         Integer no,
-        Double basePrice,
-        Double addOnPrice,
+        Integer basePrice,
+        Integer addOnPrice,
         Integer capacity,
         Floor floor,
         RoomStatus currentStatus
